refactor(core): refactored core to smaller sub folders and updated imports
no logic changes
