--- src/core/state.py
@@ -1,1 +1,0 @@
-from core.dataset_state import DatasetState  # noqa: F401
