+++ src/core/utils/geometry.py
