+++ src/models/dataset_model.py
@@ -5,8 +5,8 @@ from pathlib import Path
 from PySide6.QtCore import Qt, QAbstractTableModel, QModelIndex
 from PySide6.QtGui import QColor, QBrush
 
-from core.dataset_state import DatasetState
-from core.utils import polygon_area
+from core.states.dataset_state import DatasetState
+from core.utils.geometry import polygon_area
 
 logger = logging.getLogger("AnnoMate.DatasetModel")
