+++ src/models/inference_model.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from core.inference_state import InferenceState
+from core.states.inference_state import InferenceState
 
 
 class InferenceModel:
