+++ src/tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from core.dataset_state import DatasetState
+from core.states.dataset_state import DatasetState
 
 
 @pytest.fixture
