+++ src/tests/unit/test_inference_state.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from core.inference_state import InferenceState
+from core.states.inference_state import InferenceState
 
 
 @pytest.fixture
